+++ src/main/java/com/klarna/hiverunner/builder/Script.java
@@ -13,18 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.klarna.hiverunner.sql.split;
+package com.klarna.hiverunner.builder;
 
-import java.util.StringTokenizer;
+import java.nio.file.Path;
 
-/**
- * Provides a means to modify and inspect the state of the parsing and splitting
- * of a script.
- */
-public interface Context {
-    StringTokenizer tokenizer();
-    String statement();
-    void append(String chars);
-    void appendWith(Consumer consumer);
-    void flush();
+public interface Script {
+
+  /**
+   * index of script within all scripts in source
+   */
+  int getIndex();
+  
+  Path getPath();
+
+  String getSql();
+
 }